overcloud: disable iscsid.socker right after installing libguestfs-tools
libguestfs-tools pulls as a dependency iscsid rpms, and enable by
default the iscsi.socket. So if you reboot the undercloud, the
iscsi.socket will be enabled and running on the host which conflicts
with the iscsi container deployed on the undercloud, used by Nova and
Ironic. This patch will make sure we disable the socket on the host.