Define ironic root_device by size
The root_device defaults to name: "sda,hda,vda" which does not seem
correct. The documentation [1] indicates this should be the (exact)
device name, e.g /dev/md0, and defines how to use operators if
collections should be used. The use of name is also not recommended
in the same documentation as it can vary.

Because the OSD(s) are 50G, the size of the root disk can be used
to differentiate it from the OSDs. This is done with Ironic's
root_device size attribute.

[1] https://docs.openstack.org/ironic/pike/install/include/root-device-hints.html