Default to Stein
Default to Stein instead of Rocky. Also update associated Ceph
version. Nautilus is tested with Stein.

Also, default the hypervisor inventory ansible_host to the
loopback device. This is slightly opinionated as it assumes
the playbook is run directly on the hypervisor however it is
more likely a user would run this on the hypervisor than on
Cedric's private IP picked out of RFC1918 class A space.