Correcting index for baremetal.json generation
Apparently there were some issues with the way the index was calculated,
and it seems to be linked to the amount of overclouds. So let's link it,
and see how it goes. At least for 1 and 2, that's good. Not sure we'll
actually have more than 2, at least on a lab... ?