Allow to slow down the lab
Using the new "iotune" option in the "vms" dict, you can ensure your lab
is slow enough, even if running on the fastest SSD on the market.

An example environment file is provided, as well as some doc updates in
the README and USAGE.

The option is per-vm, meaning you can affect only a subset of instances,
like "only the undercloud", "only controller-1" and so on.

The tuning also affects the swap in case you activate it. This makes the
overall system really slow, allowing to ensure no race condition nor
weird timeouts are reached.

The provided values in the example are probably a bit too extrem, since
disk writes are limited to max 20M/s. The constant load is at about 4.