Correct and improve image-prepare and container-image-modify
Previous patch was working (although USAGE example was wrong), but it
wasn't that good.

Now, we generate the "containers-prepare-parameter.yaml" using the
original CLI call, and add to it any options, using the layering support
as described in the upstream doc[1]

This means that all the options related to container images are now
deprecated, and you should use the modify_image parameter in order to
override, for example, ceph version, or container registry.

[1]
https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/deployment/container_image_prepare.html#layering-image-preparation-entries