Example playbook building the container images on nat-vm
This example shows how you can leverage the power of tripleo-lab with
tripleo-operator - this allows to go far, far beyond the lab itself.

In this playbook, we reconfigure the nat-vm in order to build the
container images, using the dedicated operator, and push them to a local
registry (not managed in the playbook).

It then goes back to the standard deploy steps, creating the undercloud,
overcloud and running as usual.

While this example mostly works, you'll have to pass at least an
environment file in order to override the container namespace and tags
on the undercloud in order to use those built containers.