Update/correct how network is configured/used
New iterations[1] of the network provisionning seem to break the
"anti-feature" used in tripleo-lab. We can't rely only on the
service_net_map_replace parameter. A future iteration on tripleo-lab
will get completely rid of that param in order to converge a bit more
toward "best practices".

Also take the opportunity to clean useless/conflicting files.

[1] https://github.com/openstack/tripleo-heat-templates/commit/ff30e794d12f9d32bdafaf78c8e67539b8df7e39