Leverage ValidateControllersIcmp parameter
Since the internal node validations have moved to ansible, it seems the
"ping" check is a bit wider than before, since it wants to ping every
single IPs associated to a controller - before, it was apparently only
on the ctlplane network.

With the way network isolation is done in the lab, the undercloud has no
access to the "External" network, at least not at that point. In order
to allow devs to still use this project, let's deactivate this
validation.

A proper correction will come soon hopefully.