Replace unarchive with get_url
unarchive src: :// is very simple [1]. If something is already in dst it
won't redownload and reunpack. The same is for bumping version in env
variable. This patch replaces to more robust way recommended in [1] in
src section. Also this patch bumps version of appliance to 1.40.0 and
changes installation order to run libguestfs.yaml after libvirt.

[1] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/unarchive_module.html