Ensure nat-vm uses the right image
In order to avoid weird situation, we have to ensure nat-vm uses the
same image as the undercloud. That's especially important when we use
the "custom_repositories" setting, since it will push those repos on
both nat-vm and undercloud (usually used in order to get a working
caching proxy, hitting always the same upstream repository).