Add the sneaky missing "g"
Having a missing g causes me discomfort. I could
have settled for a ' to denote that we all understand
there should have been a g, but the g has been deliberately
dropped. Maybe in an effort to show that ssh-agent is cool, and
runnin'. But without a g or a ' its all just too much for me.