Enforce BIOS mode for the overcloud boot mode
A recent switch in master moved the boot_mode from bios to uefi. This
prevented the provisioning of the OC nodes, since they are, by default,
in BIOS mode.

This patch enforces BIOS mode (for now) and adds some more specific logs
for the OC virtual machines.

A follow-up is already being worked on, in order to actually get UEFI
VMs, but until now, it was blocked by secure-boot and the unability to
disable it. Some more investigations are needed on that field.