Put 'excludes' under ContainerImagePrepare
The last two lines of containers-prepare-parameter.yaml
genereated after c950d47f7db9eadc75dc61509e1439de309fcf9d
merged has "DockerInsecureRegistryAddress: []" and then
"excludes: []". However, this is not valid for "excludes"
as it is not a THT parameter. As per the tripleo-docs file
container_image_prepare.rst, the excludes list should be
in ContainerImagePrepare but not under "set" like the
tripleo-lab variable container_prepare_overrides.

This patch removes "excludes" from the jinja2 template
and updates the "fcip" tripleo-lab variable to contain
"container_prepare_excludes". This is done before "fcip"
is updated to "set" what is in the tripleo-lab variable
"container_prepare_overrides".