Start EDPM support
This patch allows to create an undercloud + any number of compute nodes
that will use the same base image (i.e. cs-9) as the undercloud. Those
nodes will also be booted, like the undercloud, and added to the
in-memory ansible inventory.
You can then deploy the Standalone scenario, and get ready for the next
steps.

Next steps will include:
- provision the compute(s)
- make them known by the initial Standalone (i.e. adopt them as
  hypervisor)